+++ src/fhir/R4R5-common/sdc-export.js
@@ -34,7 +34,7 @@ var self = {
    *  wish to put all of the returned resources into a transaction Bundle for
    *  creating them on a FHIR server.
    */
-  convertLFormsToQRAndExtracFHIRData: function (lfData, noExtensions, subject) {
+  convertLFormsToQRAndExtractFHIRData: function (lfData, noExtensions, subject) {
     var qr = this.convertLFormsToQuestionnaireResponse(lfData, noExtensions, subject);
     if (!qr.id) {
       qr.id = this._commonExport._getUniqueId(
@@ -47,6 +47,17 @@ var self = {
     for (var i = 0, len = lfData.itemList.length; i < len; ++i) {
       var item = lfData.itemList[i];
       if (this._getExtractValue(item) && this._hasItemValue(item)) {
+        const categCodeableConcepts = [];
+        var ancestor = item;
+        while (ancestor && !categCodeableConcepts.length) {
+          if (ancestor.extension) {
+            const categExts = LForms.Util.findObjectInArray(ancestor.extension, 'url',
+              this.fhirExtObsExtractCategory,  0, true);
+            categExts.forEach((x)=>categCodeableConcepts.push(x.valueCodeableConcept));
+          }
+          ancestor = ancestor._parentItem;
+        }
+
         var obs = this._commonExport._createObservation(item);
         for (var j = 0, jLen = obs.length; j < jLen; j++) {
           // Following
@@ -61,6 +72,8 @@ var self = {
           }
           if (qr.author) obs[j].performer = qr.author;
           obs[j].derivedFrom = [{ reference: qrRef }];
+          if (categCodeableConcepts.length)
+            obs[j].category = categCodeableConcepts;
 
           rtn.push(obs[j]);
         }
@@ -120,7 +133,7 @@ var self = {
     // http://hl7.org/fhir/StructureDefinition/regex
     // http://hl7.org/fhir/StructureDefinition/minValue
     // http://hl7.org/fhir/StructureDefinition/maxValue
-    // http://hl7.org/fhir/StructureDefinition/maxDecimalPlaces, not supported yet
+    // http://hl7.org/fhir/StructureDefinition/maxDecimalPlaces
     // http://hl7.org/fhir/StructureDefinition/maxSize, for attachment, not supported yet
     // maxLength
     if (item.restrictions) {
@@ -181,6 +194,17 @@ var self = {
               extValue = {
                 url: "http://hl7.org/fhir/StructureDefinition/regex",
                 valueString: value,
+              };
+            }
+            break;
+          // http://hl7.org/fhir/StructureDefinition/maxDecimalPlaces
+          case "maxDecimalPlaces":
+            if (dataType === "REAL") {
+              extValue = {
+                // You can't use self.fhirExtUrlMaxDecimalPlaces. The self object won't have
+                // the property because it goes through Object.assign() in fhirRequire.js.
+                url: this.fhirExtUrlMaxDecimalPlaces,
+                valueInteger: value,
               };
             }
             break;