+++ src/fhir/common/sdc-import-common.js
@@ -1,18 +1,6 @@
 import { InternalUtil } from '../../lib/lforms/internal-utils.js';
-import {importFHIRQuantity} from './import-common.js'
-
-// TBD import this path function from fhirpath.js.  When that is done, also
-// remove the regex test for /Quantity$/ below and replace it with a simple
-// equality check for a path of 'Quantity'.
-/**
- *  For a given result of a fhirpath.js evaluation, returns the path from the
- *  nearest FHIR type to the result which might be a fragement of that type.
- *  (Example:  Questionnaire.item, given a result consisting of items.)
- */
-function path(fhirpathRes) {
-  return fhirpathRes.__path__;
-}
-
+import {importFHIRQuantity} from './import-common.js';
+const fhirpath = require('fhirpath');
 
 /**
  *  Defines SDC import functions that are the same across the different FHIR
@@ -48,6 +36,8 @@ function addCommonSDCImportFns(ns) {
   self.fhirExtInitialExp = "http://hl7.org/fhir/uv/sdc/StructureDefinition/sdc-questionnaire-initialExpression";
   self.fhirExtObsLinkPeriod = "http://hl7.org/fhir/uv/sdc/StructureDefinition/sdc-questionnaire-observationLinkPeriod";
   self.fhirExtObsExtract = 'http://hl7.org/fhir/uv/sdc/StructureDefinition/sdc-questionnaire-observationExtract';
+  self.fhirExtObsExtractCategory =
+    "http://hl7.org/fhir/uv/sdc/StructureDefinition/sdc-questionnaire-observation-extract-category";
   self.fhirExtAnswerExp = "http://hl7.org/fhir/uv/sdc/StructureDefinition/sdc-questionnaire-answerExpression";
   self.fhirExtEnableWhenExp = "http://hl7.org/fhir/uv/sdc/StructureDefinition/sdc-questionnaire-enableWhenExpression";
   self.fhirExtChoiceOrientation = "http://hl7.org/fhir/StructureDefinition/questionnaire-choiceOrientation";
@@ -57,6 +47,7 @@ function addCommonSDCImportFns(ns) {
   self.fhirExtUnitOpen = "http://hl7.org/fhir/uv/sdc/StructureDefinition/sdc-questionnaire-unitOpen";
   self.fhirExtUnitSuppSystem = "http://hl7.org/fhir/uv/sdc/StructureDefinition/sdc-questionnaire-unitSupplementalSystem";
   self.fhirExtEntryFormat = "http://hl7.org/fhir/StructureDefinition/entryFormat";
+  self.fhirExtUrlMaxDecimalPlaces = "http://hl7.org/fhir/StructureDefinition/maxDecimalPlaces";
   self.fhirExtUrlOptionScore_lookup = {
     'STU3': "http://hl7.org/fhir/StructureDefinition/questionnaire-ordinalValue",
     'R4': "http://hl7.org/fhir/StructureDefinition/ordinalValue",
@@ -68,7 +59,8 @@ function addCommonSDCImportFns(ns) {
     self.fhirExtUrlMinValue,
     self.fhirExtUrlMaxValue,
     self.fhirExtUrlMinLength,
-    self.fhirExtUrlRegex
+    self.fhirExtUrlRegex,
+    self.fhirExtUrlMaxDecimalPlaces
   ];
 
   // One way or the other, the following extensions are converted to lforms internal fields.
@@ -90,7 +82,8 @@ function addCommonSDCImportFns(ns) {
     self.fhirExtUrlHidden,
     self.fhirExtTerminologyServer,
     self.fhirExtUrlDataControl,
-    self.fhirExtChoiceOrientation
+    self.fhirExtChoiceOrientation,
+    self.fhirExtUrlMaxDecimalPlaces
   ]);
 
   // Simple functions for mapping extensions to properties in the internal structure.
@@ -178,8 +171,10 @@ function addCommonSDCImportFns(ns) {
     'url'
   ];
 
+  // Item-level fields that are simply copied from the FHIR Questionnaire format to the LHC-Forms format, and back.
   self.itemLevelIgnoredFields = [
-    'definition'
+    'definition',
+    'id'
   ];
 
   /**
@@ -422,11 +417,8 @@ function addCommonSDCImportFns(ns) {
     var lfDataType = lfItem.dataType;
     var answers = [];
     const messages = [];
-    const fhirValPath = path(fhirVals); // TBD - should be on each value, as they might vary
     for (let i=0, len=fhirVals.length; i<len; ++i) {
       let fhirVal = fhirVals[i];
-      if (typeof fhirVal == 'object')
-        fhirVal.__path__ = fhirValPath; // TBD - work around for getting path on individual nodes
       var answer = undefined; // reset back to undefined each iteration
       let errors = {};
       let hasMessages = false;
@@ -477,8 +469,7 @@ function addCommonSDCImportFns(ns) {
       }
       else {
         if((lfDataType === 'QTY' || lfDataType === 'REAL' || lfDataType === 'INT') &&
-            (fhirVal._type === 'Quantity' || /Quantity$/.test(path(fhirVal)))) {
-          delete fhirVal.__path__;
+            (fhirVal._type === 'Quantity' || fhirpath.types(fhirVal)[0] === 'FHIR.Quantity')) {
           fhirVal._type = 'Quantity';
           [answer, errors] = this._convertFHIRQuantity(lfItem, fhirVal);
           hasMessages = !!errors;
@@ -1320,7 +1311,10 @@ function addCommonSDCImportFns(ns) {
             }));
           }
           else { // use FHIR context
-            var fhirClient = LForms.fhirContext.client;
+            var fhirClient = LForms.fhirContext?.client;
+            if (!fhirClient) {
+              throw new Error("Unable to load ValueSet "+item.answerValueSet+ " from FHIR server");
+            }
             pendingPromises.push(fhirClient.request({
                 url: lfData._buildURL(
                   ['ValueSet', '$expand'], {url: item.answerValueSet, _format: 'json'}),
@@ -1656,6 +1650,9 @@ function addCommonSDCImportFns(ns) {
         }
         else if(restriction.url.match(/regex$/)) {
           restrictions['pattern'] = val;
+        }
+        else if(restriction.url.match(/maxDecimalPlaces$/)) {
+          restrictions['maxDecimalPlaces'] = parseInt(val);
         }
       }
     }